requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==3.2.13
django-environ==0.8.1
Pillow==9.1.1
psycopg2-binary==2.9.3
sqlparse==0.4.2
